+++ transforms.h
@@ -1,9 +1,9 @@
 #ifndef _TRANSFORMS_H_
 #define _TRANSFORMS_H_
 
-void transforms_park(int alpha, int beta, int theta, int *d, int *q);
-void transforms_inverse_park(int d, int q, int theta, int *alpha, int *beta);
-void transforms_clarke(int a, int b, int c, int *alpha, int *beta);
-void transforms_inverse_clarke(int alpha, int beta, int *a, int *b, int *c);
+void transforms_park(float alpha, float beta, float theta, float *d, float *q);
+void transforms_inverse_park(float d, float q, float theta, float *alpha, float *beta);
+void transforms_clarke(float a, float b, float c, float *alpha, float *beta);
+void transforms_inverse_clarke(float alpha, float beta, float *a, float *b, float *c);
 
 #endif
